requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.4.1
motor==3.5.0
pydantic==2.5.2
pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
numpy==1.26.2
python-multipart==0.0.6
pymongo==4.6.0
